21/04/2019

+++ src/vn/edu/vnuk/record/mvc/dao/ContactDao.java
@@ -101,7 +101,7 @@ public class ContactDao {
 
     //  READ (Single Contact)
     @SuppressWarnings("finally")
-    public Contact read(int id) throws SQLException{
+    public Contact read(Long id) throws SQLException{
 
         String sqlQuery = "select * from contacts where id=?";
